Warnings cleanup
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@6505 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/InteractionsProcessor.java
@@ -595,8 +595,8 @@ public class InteractionsProcessor<T extends PreparationsProviderIFace>
                    {
                        for (Object[] row : rows)
                        {
-                           int qty    = getInt((Integer)row[1]);
-                           int qtyRes = getInt((Integer)row[2]);
+                           int qty    = getInt(row[1]);
+                           int qtyRes = getInt(row[2]);
                            prepIdToLoanQnt.put((Integer)row[0], qty-qtyRes);
                        }
                    }
